feat: be truly dict-like.
Inheriting from UserDict is kindof ugly.  But consider it to be called
'DictLike': Button()s *are* dict-like in this implementation, and it makes sense
for them to have dict-like properties (like .items()).

The alternative is to move the hooks into a custom object and compose, which is
neater *except* that `__setitem__` needs access to the long timer.  The
alternative is to have whatever triggers long_timer *first* check for a new fn,
which is even uglier.